CHanged Weapon Auto Generator to else if statement. Created dictionary<int, weapondataso>
to represent possible weapons for the individual enemies. This is set and passed into the attack states. the interfaces check the current attack in enemysodata and return the matching weapondataso and then send it to the weapon for use. Deriving enemy classes can access the enemySOData and increment up or down the current attack integer to change what attack they want to use.