BattleArenaManager, BattleArena, BattleArenaDataSO scripts all created. Added artifactInventory to saved Data in saveManager.
Future implementation will need scene handling on win and loss, reward item spawn on win. Could look like GameManager tracking what state the game is in and having the scene manager load the corresponding scene after checking the game manager . Also changed the Partner to initialize as idle if bool is checked on that instance in the inspector.