Fixed UI navigation EventSystem
to change first selected and current selected game object to be dynamic. This includes menu panels (artifact and Inventory) setting which button to be on when enabled. This also includes dialogue response buttons to set themselves as the current and first selected when they are enabled. Right now, so long as menu is enabled appropriately, it will set its button to be the one selected.